+++ base/tests/test_dna_utils.py
@@ -13,6 +13,7 @@ import os
 import unittest
 
 from base.dna_utils import DnaUtil
+from base.file_utils import FileUtil
 from base.tests.fixtures import FIXTURE_PATH
 
 GOOD_DNA = 'ACGTACGT'
@@ -21,10 +22,22 @@ BAD_DNA = 'ACGBT2'
 TEST_FWD_SEQ = 'AATTGGCC'
 TEST_REV_COMP_SEQ = 'GGCCAATT'
 
-FIXTURE1_FNAME = 'SamplePatternCount.txt'
-FIXTURE2_FNAME = 'PatternCount.txt'
+FIXTURE1_FNAME = 'SampleReverseComplement.txt'
+FIXTURE2_FNAME = 'ReverseComplement.txt'
+FIXTURE3_FNAME = 'ReverseComplementTest1.txt'
+FIXTURE4_FNAME = 'SampleLocatePattern.txt'
+FIXTURE5_FNAME = 'LocatePattern1.txt'
+FIXTURE6_FNAME = 'LocatePattern2.txt'
+FIXTURE7_FNAME = 'LocatePattern3.txt'
+FIXTURE8_FNAME = 'LocatePattern4.txt'
 FIXTURE1_FPATH = os.path.join(FIXTURE_PATH, FIXTURE1_FNAME)
 FIXTURE2_FPATH = os.path.join(FIXTURE_PATH, FIXTURE2_FNAME)
+FIXTURE3_FPATH = os.path.join(FIXTURE_PATH, FIXTURE3_FNAME)
+FIXTURE4_FPATH = os.path.join(FIXTURE_PATH, FIXTURE4_FNAME)
+FIXTURE5_FPATH = os.path.join(FIXTURE_PATH, FIXTURE5_FNAME)
+FIXTURE6_FPATH = os.path.join(FIXTURE_PATH, FIXTURE6_FNAME)
+FIXTURE7_FPATH = os.path.join(FIXTURE_PATH, FIXTURE7_FNAME)
+FIXTURE8_FPATH = os.path.join(FIXTURE_PATH, FIXTURE8_FNAME)
 
 
 class TestDnaUtilsMethods(unittest.TestCase):
@@ -68,4 +81,92 @@ class TestDnaUtilsMethods(unittest.TestCase):
         dna_util_obj = DnaUtil(TEST_FWD_SEQ)
         expected = TEST_REV_COMP_SEQ
         actual = dna_util_obj.reverse_complement()
+        self.assertEqual(expected, actual)
+
+    def test_reverse_complement_from_sample_dataset(self):
+        """
+        Test reverse_complement method with 'Sample' dataset
+        """
+        file_util_obj = FileUtil(FIXTURE1_FPATH)
+        inputs = file_util_obj.inputs
+        expected = file_util_obj.outputs[0]
+        sequence_obj = DnaUtil(inputs[0])
+        actual = sequence_obj.reverse_complement()
+        self.assertEqual(expected, actual)
+
+    def test_reverse_complement_from_extra_dataset(self):
+        """
+        Test reverse_complement method with 'Extra' dataset
+        """
+        file_util_obj = FileUtil(FIXTURE2_FPATH)
+        inputs = file_util_obj.inputs
+        expected = file_util_obj.outputs[0]
+        sequence_obj = DnaUtil(inputs[0])
+        actual = sequence_obj.reverse_complement()
+        self.assertEqual(expected, actual)
+
+    def test_reverse_complement_from_test_dataset_one(self):
+        """
+        Test reverse_complement method with 'Test' dataset
+        """
+        file_util_obj = FileUtil(FIXTURE3_FPATH)
+        inputs = file_util_obj.inputs
+        expected = file_util_obj.outputs[0]
+        sequence_obj = DnaUtil(inputs[0])
+        actual = sequence_obj.reverse_complement()
+        self.assertEqual(expected, actual)
+
+    def test_locate_pattern_from_sample_dataset(self):
+        """
+        Test locate_pattern method with 'Sample' dataset
+        """
+        file_util_obj = FileUtil(FIXTURE4_FPATH)
+        inputs = file_util_obj.inputs
+        expected = [int(i) for i in file_util_obj.outputs]
+        sequence_obj = DnaUtil(inputs[1])
+        actual = sequence_obj.locate_pattern_in_sequence(inputs[0])
+        self.assertEqual(expected, actual)
+
+    def test_locate_pattern_from_dataset_one(self):
+        """
+        Test locate_pattern method with 'Test' dataset one
+        """
+        file_util_obj = FileUtil(FIXTURE5_FPATH)
+        inputs = file_util_obj.inputs
+        expected = [int(i) for i in file_util_obj.outputs]
+        sequence_obj = DnaUtil(inputs[1])
+        actual = sequence_obj.locate_pattern_in_sequence(inputs[0])
+        self.assertEqual(expected, actual)
+
+    def test_locate_pattern_from_dataset_two(self):
+        """
+        Test locate_pattern method with 'Test' dataset two
+        """
+        file_util_obj = FileUtil(FIXTURE6_FPATH)
+        inputs = file_util_obj.inputs
+        expected = [int(i) for i in file_util_obj.outputs]
+        sequence_obj = DnaUtil(inputs[1])
+        actual = sequence_obj.locate_pattern_in_sequence(inputs[0])
+        self.assertEqual(expected, actual)
+
+    def test_locate_pattern_from_dataset_three(self):
+        """
+        Test locate_pattern method with 'Test' dataset three
+        """
+        file_util_obj = FileUtil(FIXTURE7_FPATH)
+        inputs = file_util_obj.inputs
+        expected = [int(i) for i in file_util_obj.outputs]
+        sequence_obj = DnaUtil(inputs[1])
+        actual = sequence_obj.locate_pattern_in_sequence(inputs[0])
+        self.assertEqual(expected, actual)
+
+    def test_locate_pattern_from_dataset_four(self):
+        """
+        Test locate_pattern method with 'Test' dataset four
+        """
+        file_util_obj = FileUtil(FIXTURE8_FPATH)
+        inputs = file_util_obj.inputs
+        expected = [int(i) for i in file_util_obj.outputs]
+        sequence_obj = DnaUtil(inputs[1])
+        actual = sequence_obj.locate_pattern_in_sequence(inputs[0])
         self.assertEqual(expected, actual)